Split and restructure code, use Meson
Making smaller units is nice for understanding the codebase and making
it extensible (e.g. to use something else but COMEDI for input).
This commit keeps the original overall structure of the codebase though.

This also switches the build system to Meson, to be easier to maintain
and faster to build.