Optimize DAQ codepath
The latency was way too high, this patch brings it down to expected
levels and also improves the code readability a bit (not a thing these
optimizations usually do).
The previous routines to get single pieces from the buffer were never
meant to be used in a hot codepath like the one acquiring actual SWR
data.